+++ src/events/client/MessageCreate.ts
@@ -11,38 +11,38 @@ export default class MessageCreate extends Event {
     // biome-ignore lint/complexity/noExcessiveCognitiveComplexity: <explanation>
     public async run(message: Message): Promise<any> {
         if (message.author.bot) return;
+
         const setup = await this.client.db.getSetup(message.guildId);
         if (setup && setup.textId === message.channelId) {
             return this.client.emit("setupSystem", message);
         }
+
         const guild = await this.client.db.get(message.guildId);
         const mention = new RegExp(`^<@!?${this.client.user.id}>( |)$`);
-        if (message.content.match(mention)) {
+        if (mention.test(message.content)) {
             await message.reply({
                 content: `Hey, my prefix for this server is \`${guild.prefix}\` Want more info? then do \`${guild.prefix}help\`\nStay Safe, Stay Awesome!`,
             });
             return;
         }
+
         const escapeRegex = (str: string): string => str.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
         const prefixRegex = new RegExp(`^(<@!?${this.client.user.id}>|${escapeRegex(guild.prefix)})\\s*`);
         if (!prefixRegex.test(message.content)) return;
+
         const [matchedPrefix] = message.content.match(prefixRegex);
         const args = message.content.slice(matchedPrefix.length).trim().split(/ +/g);
         const cmd = args.shift()?.toLowerCase();
         const command = this.client.commands.get(cmd) || this.client.commands.get(this.client.aliases.get(cmd) as string);
         if (!command) return;
+
         const ctx = new Context(message, args);
         ctx.setArgs(args);
-        let dm = message.author.dmChannel;
-        if (typeof dm === "undefined") dm = await message.author.createDM();
-        if (
-            !(
-                message.inGuild() &&
-                message.channel.permissionsFor(message.guild.members.resolve(this.client.user))?.has(PermissionFlagsBits.ViewChannel)
-            )
-        )
-            return;
-        if (!message.guild.members.resolve(this.client.user).permissions.has(PermissionFlagsBits.SendMessages)) {
+
+        if (!message.guild.members.resolve(this.client.user)?.permissions.has(PermissionFlagsBits.ViewChannel)) return;
+
+        const clientMember = message.guild.members.resolve(this.client.user);
+        if (!clientMember.permissions.has(PermissionFlagsBits.SendMessages)) {
             await message.author
                 .send({
                     content: `I don't have **\`SendMessage\`** permission in \`${message.guild.name}\`\nchannel: <#${message.channelId}>`,
@@ -50,33 +50,35 @@ export default class MessageCreate extends Event {
                 .catch(() => {});
             return;
         }
-        if (!message.guild.members.resolve(this.client.user).permissions.has(PermissionFlagsBits.EmbedLinks)) {
+
+        if (!clientMember.permissions.has(PermissionFlagsBits.EmbedLinks)) {
             await message.reply({
                 content: "I don't have **`EmbedLinks`** permission.",
             });
             return;
         }
+
         if (command.permissions) {
-            if (
-                command.permissions.client &&
-                !message.guild.members.resolve(this.client.user).permissions.has(command.permissions.client)
-            ) {
+            if (command.permissions.client && !clientMember.permissions.has(command.permissions.client)) {
                 await message.reply({
                     content: "I don't have enough permissions to execute this command.",
                 });
                 return;
             }
+
             if (command.permissions.user && !message.member.permissions.has(command.permissions.user)) {
                 await message.reply({
                     content: "You don't have enough permissions to use this command.",
                 });
                 return;
             }
+
             if (command.permissions.dev && this.client.config.owners) {
-                const findDev = this.client.config.owners.find((x) => x === message.author.id);
-                if (!findDev) return;
+                const isDev = this.client.config.owners.includes(message.author.id);
+                if (!isDev) return;
             }
         }
+
         if (command.player) {
             if (command.player.voice) {
                 if (!message.member.voice.channel) {
@@ -85,48 +87,49 @@ export default class MessageCreate extends Event {
                     });
                     return;
                 }
-                if (!message.guild.members.resolve(this.client.user).permissions.has(PermissionFlagsBits.Speak)) {
+
+                if (!clientMember.permissions.has(PermissionFlagsBits.Connect)) {
                     await message.reply({
                         content: `I don't have \`CONNECT\` permissions to execute this \`${command.name}\` command.`,
                     });
                     return;
                 }
-                if (!message.guild.members.resolve(this.client.user).permissions.has(PermissionFlagsBits.Speak)) {
+
+                if (!clientMember.permissions.has(PermissionFlagsBits.Speak)) {
                     await message.reply({
                         content: `I don't have \`SPEAK\` permissions to execute this \`${command.name}\` command.`,
                     });
                     return;
                 }
+
                 if (
                     message.member.voice.channel.type === ChannelType.GuildStageVoice &&
-                    !message.guild.members.resolve(this.client.user).permissions.has(PermissionFlagsBits.RequestToSpeak)
+                    !clientMember.permissions.has(PermissionFlagsBits.RequestToSpeak)
                 ) {
                     await message.reply({
                         content: `I don't have \`REQUEST TO SPEAK\` permission to execute this \`${command.name}\` command.`,
                     });
                     return;
                 }
-                if (
-                    message.guild.members.resolve(this.client.user).voice.channel &&
-                    message.guild.members.resolve(this.client.user).voice.channelId !== message.member.voice.channelId
-                ) {
+
+                if (clientMember.voice.channel && clientMember.voice.channelId !== message.member.voice.channelId) {
                     await message.reply({
-                        content: `You are not connected to <#${
-                            message.guild.members.resolve(this.client.user).voice.channel.id
-                        }> to use this \`${command.name}\` command.`,
+                        content: `You are not connected to <#${clientMember.voice.channelId}> to use this \`${command.name}\` command.`,
                     });
                     return;
                 }
             }
+
             if (command.player.active) {
                 const queue = this.client.queue.get(message.guildId);
-                if (!queue?.queue && queue.current) {
+                if (!(queue?.queue && !queue.current)) {
                     await message.reply({
                         content: "Nothing is playing right now.",
                     });
                     return;
                 }
             }
+
             if (command.player.dj) {
                 const dj = await this.client.db.getDj(message.guildId);
                 if (dj?.mode) {
@@ -137,20 +140,20 @@ export default class MessageCreate extends Event {
                         });
                         return;
                     }
-                    const findDJRole = message.member.roles.cache.find((x: any) => djRole.map((y: any) => y.roleId).includes(x.id));
-                    if (!findDJRole) {
-                        if (!message.member.permissions.has(PermissionFlagsBits.ManageGuild)) {
-                            await message
-                                .reply({
-                                    content: "You need to have the DJ role to use this command.",
-                                })
-                                .then((msg) => setTimeout(() => msg.delete(), 5000));
-                            return;
-                        }
+
+                    const hasDJRole = message.member.roles.cache.some((role) => djRole.map((r) => r.roleId).includes(role.id));
+                    if (!(hasDJRole && !message.member.permissions.has(PermissionFlagsBits.ManageGuild))) {
+                        await message
+                            .reply({
+                                content: "You need to have the DJ role to use this command.",
+                            })
+                            .then((msg) => setTimeout(() => msg.delete(), 5000));
+                        return;
                     }
                 }
             }
         }
+
         if (command.args && !args.length) {
             const embed = this.client
                 .embed()
@@ -165,12 +168,15 @@ export default class MessageCreate extends Event {
             await message.reply({ embeds: [embed] });
             return;
         }
+
         if (!this.client.cooldown.has(cmd)) {
             this.client.cooldown.set(cmd, new Collection());
         }
+
         const now = Date.now();
         const timestamps = this.client.cooldown.get(cmd)!;
-        const cooldownAmount = Math.floor(command.cooldown || 5) * 1000;
+        const cooldownAmount = (command.cooldown || 5) * 1000;
+
         if (timestamps.has(message.author.id)) {
             const expirationTime = timestamps.get(message.author.id)! + cooldownAmount;
             const timeLeft = (expirationTime - now) / 1000;
@@ -186,21 +192,23 @@ export default class MessageCreate extends Event {
             timestamps.set(message.author.id, now);
             setTimeout(() => timestamps.delete(message.author.id), cooldownAmount);
         }
+
         if (args.includes("@everyone") || args.includes("@here")) {
             await message.reply({
                 content: "You can't use this command with everyone or here.",
             });
             return;
         }
+
         try {
             return command.run(this.client, ctx, ctx.args);
         } catch (error) {
             this.client.logger.error(error);
             await message.reply({ content: `An error occurred: \`${error}\`` });
-            return;
         }
     }
 }
+
 /**
  * Project: lavamusic
  * Author: Appu
